requirements.txt
# Core Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Manipulation & Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.2.0

# Visualization & UI
matplotlib>=3.7.0
tqdm>=4.65.0

# Optional: For GPU acceleration on Windows/Linux
# Ensure you have the matching CUDA Toolkit installed on your system.